% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp.R
\name{gcp_deletejob}
\alias{gcp_deletejob}
\title{Delete job}
\usage{
gcp_deletejob(jobid)
}
\arguments{
\item{jobid}{ID of job to delete}
}
\description{
Delete job
}
\note{
For info about parameters see \url{https://developers.google.com/cloud-print/docs/appInterfaces#deletejob}
}
