% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp.R
\name{gcp_jobs}
\alias{gcp_jobs}
\title{Query submitted jobs}
\usage{
gcp_jobs(printerid = NULL, owner = NULL, status = NULL, query = NULL,
  offset = NULL, limit = NULL, sortorder = NULL)
}
\arguments{
\item{printerid}{The ID of the printer}

\item{owner}{Query job owner}

\item{status}{Query job status}

\item{query}{Query title and tags}

\item{offset}{Start search at given offset}

\item{limit}{Restrict number of jobs returned}

\item{sortorder}{Choose paramter to sort by}
}
\description{
Query submitted jobs
}
\note{
For info about parameters see \url{https://developers.google.com/cloud-print/docs/appInterfaces#jobs}
}
