% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp.R
\name{gcp_processinvite}
\alias{gcp_processinvite}
\title{Accept printer invitation}
\usage{
gcp_processinvite(printerid, accept = TRUE)
}
\arguments{
\item{printerid}{The ID of the printer}

\item{accept}{Boolean indicating whether to accept request from printer}
}
\description{
Accept printer invitation
}
\note{
Undocumented API discovered at \url{https://stackoverflow.com/a/36366114/4564432}
}
