% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp.R
\name{gcp_search}
\alias{gcp_search}
\title{List printers associated with Google account}
\usage{
gcp_search(name = NULL, type = NULL, connection_status = NULL,
  use_cdd = NULL, extra_fields = NULL)
}
\arguments{
\item{name}{String for searching printer names}
}
\description{
List printers associated with Google account
}
